Implement `onClickOut` directive
This includes a solution for closing the Modal on backdrop click.
The solution constitutes of a new reusable `vClickOut` directive
which could be used in any other case where watching for a click
outside of a DOM element is needed.

Tests provided for the directive itself as well as extending the
Modal tests so they also cover the ability to control the active
state of the directive via props.

Small refactor of the Modal tests was done in order to cut down
code length and provide better readability.